Emit dragover event before drop
With real browsers, `dragover` events are issued before the final `drop`
(simply because the mouse passes over the drop area). Some drag'n'drop
libraries rely on this to allow / disallow drop to certain places.

This patch adds this functionality, so these libs work OK when used with
html-dnd.

Kudos to John Reynolds for his first implementation at
https://github.com/PloughingAByteField/html-dnd/commit/31872b

Closes: #14